perf(esmfold2): run the triangular contraction in bf16
O(L^3 * D), the largest GEMM in a fold -- about a third of all arithmetic at
L=500. The port copied esm's unfused reference path, which upcasts to fp32;
esm's fused Triton path keeps it bf16. Apple GPUs have no fp32 matmul
acceleration, so fp32 here costs roughly 2x for no reason.

+++ mlx_lm/models/esmfold2.py
@@ -106,8 +106,9 @@ class TriangleMultiplicativeUpdate(nn.Module):
         routed = signal * mx.sigmoid(gate_logits)
         if mask is not None:
             routed = routed * mask[..., None]
-        # Contraction in fp32 (matches the reference).
-        left, right = mx.split(routed.astype(mx.float32), 2, axis=-1)
+        # bf16 contraction, as esm's fused Triton path does. (esm's unfused
+        # reference path upcasts to fp32; the fused one does not.)
+        left, right = mx.split(routed.astype(mx.bfloat16), 2, axis=-1)
         contracted = self._contract(left, right).astype(z.dtype)
         mixed = self.proj_emit(self.norm_mix(contracted))
         out_gate = mx.sigmoid(self.proj_gate(normalized))
